WIP reset board popup
